+++ Gemfile
@@ -1,5 +1,2 @@
-source 'https://rubygems.org'
 
-group :jekyll_plugins do
-  gem 'jekyll', '~> 3.8.1'
-end
+gem "jekyll", "~> 4.0"
