+++ _config.yml
@@ -34,10 +34,16 @@ exclude:
   - .sass-cache/
   - js/
   - vendor/
-  - node_modules
-  - README.md
+  - .jekyll-cache/
+  - gemfiles/
   - Gemfile
   - Gemfile.lock
+  - node_modules/
+  - vendor/bundle/
+  - vendor/cache/
+  - vendor/gems/
+  - vendor/ruby/
+  - README.md
   - package.json
   - package-lock.json
   - template.hbs
