+++ assets/scss/_variables.scss
@@ -1,7 +1,7 @@
 @import "color";
 
 
-$primary:          $color-primary-8;
+$primary:          purple;
 $info:             $color-info-7;
 $success:          $color-success-7;
 $warning:          $color-warning-7;
