+++ gulpfile.js
@@ -1,116 +1,63 @@
-var gulp = require("gulp");
-var browserSync = require("browser-sync");
-var sass = require("gulp-sass");
-var minifyCSS = require("gulp-clean-css");
-var uglify = require("gulp-uglify");
-var concat = require("gulp-concat");
-var notify = require("gulp-notify");
-var prefix = require("gulp-autoprefixer");
-var cp = require("child_process");
-var fs = require("fs");
-var ghPages = require("gulp-gh-pages");
-var jekyll = process.platform === "win32" ? "jekyll.bat" : "jekyll";
-var messages = {
-  jekyllBuild: '<span style="color: gray">Running:</span> $ jekyll build'
+var gulp = require('gulp');
+var sass = require('gulp-sass');
+var cssnano = require('cssnano');
+var autoprefixer = require('autoprefixer');
+var postcss = require('gulp-postcss');
+var cp = require('child_process');
+var jekyll = process.platform === 'win32' ? 'jekyll.bat' : 'jekyll';
+var browserSync = require('browser-sync').create();
+
+var paths = {
+  styles: {
+    src: './assets/scss/documentation.css',
+    dest: './dist/assets/scss',
+  },
+  scripts: {
+    src: 'assets/js/*.js',
+    dest: 'dist/assets/js/'
+  }
 };
 
-/**
- * Build the Jekyll Site
- */
-gulp.task("jekyll-build", function(done) {
-  browserSync.notify(messages.jekyllBuild);
-  return cp
-    .spawn(jekyll, ["build", "--config", "_config.yml"], {
-      stdio: "inherit"
-    })
-    .on("close", done);
-});
+function jekyllBuild() {
+  return cp.spawn( jekyll , ['build'], {stdio: 'inherit'})
+}
 
-/**
- * Rebuild Jekyll & do page reload
- */
-gulp.task("jekyll-rebuild", ["build-html"], function() {
-  browserSync.reload();
-});
+function style() {
+  return gulp.src(paths.styles.src)
+    .pipe(sass({
+      outputStyle: 'expanded',
+      onError: browserSync.notify
+    }))
+    .pipe(postcss([
+      autoprefixer()
+    ]))
+    .pipe(gulp.dest(paths.styles.dest))
+    .pipe(browserSync.reload({stream:true}));
+}
 
-/**
- * Wait for jekyll-build, then launch the Server
- */
-gulp.task("browser-sync", ["build-html"], function() {
-  browserSync({
+function js() {
+  return gulp.src(paths.scripts.src)
+  .pipe(gulp.dest(paths.scripts.dest))
+  .pipe(browserSync.reload({stream:true}))
+}
+
+function browserSyncServe(done) {
+  browserSync.init({
     server: {
       baseDir: "dist"
     }
-  });
-});
+  })
+  done();
+}
 
-/**
- * Compile files from _scss into both dist/css (for live injecting) and site (for future jekyll builds)
- */
-gulp.task("compile-css", ["jekyll-build"], function() {
-  return gulp
-    .src("./assets/scss/documentation.css")
-    .pipe(
-      prefix(["last 15 versions", "> 1%", "ie 8", "ie 7"], {
-        cascade: true
-      })
-    )
-    .pipe(gulp.dest("dist/assets/scss/"))
-    .pipe(
-      browserSync.reload({
-        stream: true
-      })
-    );
-});
+function browserSyncReload(done) {
+  browserSync.reload();
+  done();
+}
 
-
-
-gulp.task("copy-cupcake", ["compile-css"], function() {
-  return gulp
-  .src([
-    "./node_modules/@ipreo/cupcake/dist/**/*",
-    "!./node_modules/@ipreo/cupcake/dist/**/*.html",
-    "!./node_modules/@ipreo/cupcake/dist/{demo-js,demo-js/**}",
-    "!./node_modules/@ipreo/cupcake/dist/{img,img/**}"
-  ])
-    .pipe(gulp.dest("./assets/cupcake/"))
-});
-
-
-gulp.task("compile-js", ["copy-cupcake"], function() {
-  return gulp;
-});
-
-gulp.task("compress-js", function() {
-  return gulp
-    .src("assets/js/plugins/*.js")
-    .pipe(concat("plugins.js"))
-    .pipe(gulp.dest("dist/assets/js/"));
-});
-
-gulp.task("build-html", ["copy-cupcake"], function() {
-  return gulp;
-});
-
-// gulp.task("get-new-cupcake", function() {
-//   return gulp
-//   .src([
-//     "./node_modules/@ipreo/cupcake/dist/**/*",
-//     "!./node_modules/@ipreo/cupcake/dist/**/*.html",
-//     "!./node_modules/@ipreo/cupcake/dist/{demo-js,demo-js/**}",
-//     "!./node_modules/@ipreo/cupcake/dist/{img,img/**}"
-//   ])
-//     .pipe(gulp.dest("./assets/cupcake/"))
-// });
-
-
-
-
-
-/**
- * Watch html/md files, run jekyll & reload BrowserSync
- */
-gulp.task("watch", function() {
+function watch() {
+  gulp.watch(paths.styles.src, style)
+  gulp.watch(paths.scripts.src, js)
   gulp.watch(
     [
       "assets/scss/*.scss",
@@ -123,33 +70,8 @@ gulp.task("watch", function() {
       "content/**/*.md",
       "content/*",
       "_posts/*"
-    ],
-    ["jekyll-rebuild"]
-  );
-});
+  ],
+  gulp.series(jekyllBuild, browserSyncReload));
+}
 
-/**
- * Default task, running just `gulp` will compile the sass,
- * compile the jekyll site, launch BrowserSync & watch files.
- */
-gulp.task("default", ["browser-sync", "watch"]);
-
-/**
- * Create production-ready website
- */
-gulp.task("jekyll-build-prod", function(done) {
-  browserSync.notify(messages.jekyllBuild);
-  return cp
-    .spawn(jekyll, ["build", "--config", "_config.yml,_config.prod.yml"], {
-      stdio: "inherit"
-    })
-    .on("close", done);
-});
-
-
-gulp.task("push-to-gh-pages", function() {
-  return gulp.src("./dist/**/*").pipe(ghPages());
-});
-
-gulp.task("deploy", ["jekyll-build-prod", "push-to-gh-pages"]);
-
+gulp.task('default', gulp.parallel(jekyllBuild, browserSyncServe, watch))
