+++ package.json
@@ -15,15 +15,17 @@
     "dev": "gulp"
   },
   "devDependencies": {
-    "browser-sync": "2.26.3",
-    "gulp": "3.9.1",
-    "gulp-autoprefixer": "6.0.0",
     "git-directory-deploy": "1.5.1",
-    "gulp-clean-css": "3.10.0",
-    "gulp-concat": "2.6.1",
     "gulp-gh-pages": "0.5.4",
-    "gulp-notify": "3.2.0",
-    "gulp-sass": "4.0.1",
-    "gulp-uglify": "3.0.1"
+    "autoprefixer": "^9.6.1",
+    "browser-sync": "^2.26.7",
+    "cssnano": "^4.1.10",
+    "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
+    "gulp-clean-css": "^4.2.0",
+    "gulp-postcss": "^8.0.0",
+    "gulp-sass": "^4.0.2",
+    "gulp-sourcemaps": "^2.6.5",
+    "vinyl-source-stream": "^2.0.0"
   }
 }
